Precisao da altura dos planos de corte alterada de float para double

+++ src/utils/AbstractGeradorDeAlturaDoPlanoDeCorteStrategy.cpp
@@ -3,7 +3,7 @@
 #include <ctime>
 using namespace simulacao::utils;
 
-AbstractGeradorDeAlturaDoPlanoDeCorteStrategy::AbstractGeradorDeAlturaDoPlanoDeCorteStrategy(float h0,float h1){
+AbstractGeradorDeAlturaDoPlanoDeCorteStrategy::AbstractGeradorDeAlturaDoPlanoDeCorteStrategy(double h0,double h1){
 	this->h0 = h0;
 	this->h1 = h1;
 	srand(time(0));
